+++ back/database/migrations/2024_06_04_040608_create_diarios_table.php
@@ -17,12 +17,12 @@ return new class extends Migration
             $table->string('medida');
             $table->unsignedBigInteger('category_insumo_id');
             $table->foreign('category_insumo_id')->references('id')->on('category_insumos');
-            $table->double('inicio',8,3);
-            $table->double('ingreso',8,3);
-            $table->double('egreso',8,3);
-            $table->double('ventas',8,3);
-            $table->double('cierre',8,3);
-            $table->double('local',8,3);
+            $table->decimal('inicio',8,3);
+            $table->decimal('ingreso',8,3);
+            $table->decimal('egreso',8,3);
+            $table->decimal('ventas',8,3);
+            $table->decimal('cierre',8,3);
+            $table->decimal('local',8,3);
             $table->timestamps();
         });
     }
